+++ jkube-kit/config/service/src/test/java/org/eclipse/jkube/kit/config/service/openshift/OpenShiftBuildServiceUtilsTest.java
@@ -15,8 +15,12 @@ package org.eclipse.jkube.kit.config.service.openshift;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 
+import io.fabric8.openshift.api.model.ImageStreamTag;
+import io.fabric8.openshift.api.model.ImageStreamTagBuilder;
 import org.eclipse.jkube.kit.build.api.assembly.ArchiverCustomizer;
 import org.eclipse.jkube.kit.build.api.assembly.JKubeBuildTarArchiver;
 import org.eclipse.jkube.kit.common.JKubeConfiguration;
@@ -43,10 +47,13 @@ import org.mockito.ArgumentCaptor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.eclipse.jkube.kit.config.service.openshift.OpenShiftBuildServiceUtils.computeS2IBuildName;
+import static org.eclipse.jkube.kit.config.service.openshift.OpenShiftBuildServiceUtils.createAdditionalTagsIfPresent;
 import static org.eclipse.jkube.kit.config.service.openshift.OpenShiftBuildServiceUtils.createBuildArchive;
 import static org.eclipse.jkube.kit.config.service.openshift.OpenShiftBuildServiceUtils.createBuildOutput;
 import static org.eclipse.jkube.kit.config.service.openshift.OpenShiftBuildServiceUtils.createBuildStrategy;
+import static org.eclipse.jkube.kit.config.service.openshift.OpenShiftBuildServiceUtils.getAdditionalTagsToCreate;
 import static org.eclipse.jkube.kit.config.service.openshift.OpenShiftBuildServiceUtils.getBuildConfigSpec;
+import static org.eclipse.jkube.kit.config.service.openshift.OpenshiftBuildService.DOCKER_IMAGE;
 import static org.junit.Assert.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
@@ -278,5 +285,83 @@ public class OpenShiftBuildServiceUtilsTest {
         .isNotSameAs(originalSpec)
         .isSameAs(buildConfig.getSpec())
         .hasFieldOrPropertyWithValue("runPolicy", "Serial");
+  }
+
+  @Test
+  public void createAdditionalTagsIfPresent_withNoAdditionalTag_shouldReturnEmptyList() {
+    // Given + When
+    List<ImageStreamTag> imageStreamTagList = createAdditionalTagsIfPresent(imageConfiguration, "ns1", null);
+
+    // Then
+    assertThat(imageStreamTagList).isEmpty();
+  }
+
+  @Test
+  public void createAdditionalTagsIfPresent_withAdditionalTags_shouldReturnNonEmptyImageStreamTagList() {
+    // Given
+    ImageConfiguration imageConfigurationWithAdditionalTags = createNewImageConfigurationWithAdditionalTags();
+
+    // When
+    List<ImageStreamTag> imageStreamTagList = createAdditionalTagsIfPresent(imageConfigurationWithAdditionalTags, "ns1", new ImageStreamTagBuilder()
+        .withNewMetadata().withName("test:t1").endMetadata()
+        .withNewImage().withDockerImageReference("foo-registry.openshift.svc:5000/test/test@sha256:1234").endImage()
+        .build());
+
+    // Then
+    assertThat(imageStreamTagList)
+        .hasSize(2)
+        .containsExactlyInAnyOrder(
+            new ImageStreamTagBuilder()
+                .withNewMetadata()
+                .withName("test:t2")
+                .withNamespace("ns1")
+                .endMetadata()
+                .withNewTag()
+                .withNewFrom()
+                .withKind(DOCKER_IMAGE)
+                .withName("foo-registry.openshift.svc:5000/test/test@sha256:1234")
+                .endFrom()
+                .endTag()
+                .withGeneration(0L)
+                .build(),
+            new ImageStreamTagBuilder()
+                .withNewMetadata()
+                .withName("test:t3")
+                .withNamespace("ns1")
+                .endMetadata()
+                .withNewTag()
+                .withNewFrom()
+                .withKind(DOCKER_IMAGE)
+                .withName("foo-registry.openshift.svc:5000/test/test@sha256:1234")
+                .endFrom()
+                .endTag()
+                .withGeneration(0L)
+                .build()
+        );
+  }
+
+  @Test
+  public void getAdditionalTagsToCreate_withNoAdditionalTag_shouldReturnEmptyList() {
+    // Given + When
+    List<String> additionalTags = getAdditionalTagsToCreate(imageConfiguration);
+
+    // Then
+    assertThat(additionalTags).isEmpty();
+  }
+
+  @Test
+  public void getAdditionalTagsToCreate_withAdditionalTags_shouldReturnExtraTagsList() {
+    // Given + When
+    List<String> additionalTags = getAdditionalTagsToCreate(createNewImageConfigurationWithAdditionalTags());
+
+    // Then
+    assertThat(additionalTags).containsExactlyInAnyOrder("t2", "t3");
+  }
+
+  private ImageConfiguration createNewImageConfigurationWithAdditionalTags() {
+    return ImageConfiguration.builder()
+        .name("test:t1")
+        .build(BuildConfiguration.builder().tags(Arrays.asList("t2", "t3")).build())
+        .build();
   }
 }